refactor: move pure briefing/preference helpers out of store.ts
Lifts serialiseMorningBriefing/serialiseWeeklyBriefing into
src/format/briefing.ts and the two preference helpers into
src/format/preferences.ts. Pure functions; behavior unchanged.
